@@ -9,43 +9,50 @@
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
-    "build:prod": "ng build --prod",
+    "build:prod": "ng build --prod --buildOptimizer",
     "deploy:hosting": "npm run build:prod && firebase deploy --only hosting",
     "deploy:functions": "npm run buil:prod && firebase deploy --only functions",
-    "deploy": "npm run build:prod && firebase deploy"
+    "deploy": "npm run build:prod && firebase deploy",
+    "deploy:test": "npm run build:prod && firebase deploy --only hosting:test"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.5",
-    "@angular/cdk": "^5.2.1",
-    "@angular/common": "^5.2.5",
-    "@angular/compiler": "^5.2.5",
-    "@angular/core": "^5.2.5",
-    "@angular/forms": "^5.2.5",
-    "@angular/http": "^5.2.5",
-    "@angular/material": "^5.2.1",
-    "@angular/platform-browser": "^5.2.5",
-    "@angular/platform-browser-dynamic": "^5.2.5",
-    "@angular/platform-server": "^5.2.5",
-    "@angular/router": "^5.2.5",
-    "angularfire2": "^5.0.0-rc.6",
+    "@angular/animations": "7.2.3",
+    "@angular/cdk": "^7.3.0",
+    "@angular/common": "7.2.3",
+    "@angular/compiler": "7.2.3",
+    "@angular/core": "7.2.3",
+    "@angular/fire": "^5.1.1",
+    "@angular/forms": "7.2.3",
+    "@angular/http": "7.2.3",
+    "@angular/material": "^7.3.0",
+    "@angular/platform-browser": "7.2.3",
+    "@angular/platform-browser-dynamic": "7.2.3",
+    "@angular/platform-server": "7.2.3",
+    "@angular/pwa": "^0.12.4",
+    "@angular/router": "7.2.3",
+    "@angular/service-worker": "7.2.3",
     "core-js": "^2.5.3",
-    "firebase": "^4.10.0",
+    "firebase": "^5.8.2",
     "hammerjs": "^2.0.8",
     "jquery": "^3.3.1",
     "raf": "^3.4.0",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.20"
+    "rxjs": "^6.4.0",
+    "tslib": "^1.9.0",
+    "zone.js": "~0.8.29",
+    "terser": "3.14.1"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.2",
-    "@angular/compiler-cli": "^5.2.5",
-    "@angular/language-service": "^5.2.5",
+    "@angular-devkit/build-angular": "~0.13.0",
+    "@angular/cli": "7.3.0",
+    "@angular/compiler-cli": "7.2.3",
+    "@angular/language-service": "7.2.3",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
     "@types/jquery": "^3.3.0",
     "@types/node": "~9.4.6",
     "codelyzer": "~4.2.1",
+    "cypress": "^3.1.5",
     "firebase-admin": "^5.9.0",
     "firebase-functions": "^0.8.1",
     "jasmine-core": "~3.1.0",
@@ -59,7 +66,7 @@
     "protractor": "~5.3.0",
     "ts-node": "~5.0.0",
     "tslint": "~5.9.1",
-    "typescript": "^2.7.2",
+    "typescript": "3.2.4",
     "webpack-bundle-analyzer": "^2.10.0"
   }
 }
